Move app-wide fixtures to conftest file
